+++ RedshiftLibrary/src/lib/linemodel/element.cpp
@@ -53,12 +53,9 @@ using namespace NSEpic;
  *defaults.
  **/
 CLineModelElement::CLineModelElement(
-    const TLineModelElementParam_ptr elementParam, Float64 maxDistanceToLine,
-    Int32 minSamplesNumberForLineFit)
-    : m_ElementParam(std::move(elementParam)),
-      m_maxDistanceToLine(maxDistanceToLine),
-      m_minSamplesNumberForLineFit(minSamplesNumberForLineFit),
-      m_OutsideLambdaRange(true), m_size(m_ElementParam->size()){};
+    const TLineModelElementParam_ptr elementParam)
+    : m_ElementParam(std::move(elementParam)), m_OutsideLambdaRange(true),
+      m_size(m_ElementParam->size()){};
 
 void TLineModelElementParam::resetFittingParams() {
   // init the fitted amplitude values and related variables
@@ -165,10 +162,15 @@ std::pair<Float64, Float64> CLineModelElement::GetContinuumAtCenterProfile(
   return std::make_pair(cont, contStd);
 }
 
-/**
- * \brief Returns the theoretical support range for the line
- **/
-void CLineModelElement::EstimateTheoreticalSupport(
+// Estimate the line index range inside the spectralaxis of the line wavelength
+// range. The wavelength range is computed as:
+//         mu +- [sigma * n_sigma_support/2 + max_offset]
+// where:
+//   mu: redshifted wavelength of line center
+//   sigma: line width (using LSF and velocity dispersion)
+//   max_offset: maximum possible offset of the line center in km/s
+// Then the index range is determined from the wavelength range.
+void CLineModelElement::EstimateSupport(
     Int32 line_index, const CSpectrumSpectralAxis &spectralAxis,
     Float64 redshift, const TFloat64Range &lambdaRange, Float64 max_offset) {
   Float64 mu = GetObservedPosition(line_index, redshift);
@@ -182,24 +184,32 @@ void CLineModelElement::EstimateTheoreticalSupport(
   Float64 winsize =
       getElementParam()->getLineProfile(line_index)->GetNSigmaSupport() * sigma;
   winsize += 2 * max_offset_angstrom;
-  TInt32Range supportRange =
+  m_range[line_index] =
       EstimateIndexRange(spectralAxis, mu, lambdaRange, winsize);
 
-  m_StartTheoretical[line_index] = supportRange.GetBegin();
-  m_EndTheoretical[line_index] = supportRange.GetEnd();
-  m_StartNoOverlap[line_index] = supportRange.GetBegin();
-  m_EndNoOverlap[line_index] = supportRange.GetEnd();
+  m_rangeNoOverlap[line_index] = m_range[line_index];
 
-  EstimateLineVisbility(line_index, spectralAxis, supportRange, mu, sigma,
+  EstimateLineVisbility(line_index, spectralAxis, mu, sigma,
                         max_offset_angstrom);
+
+  // if the amplitude offset (ie the polynomial under the line) is fitted, we
+  // extend the line support with a margin on both left and right sides.
+  // The added margin on each side is nSigmaAmpOffsets * sigma
+  if (getElementParam()->m_useAmpOffsetsCoeffs &&
+      !m_OutsideLambdaRangeList[line_index]) {
+    winsize += getElementParam()->m_nSigmaAmpOffsets * sigma;
+    m_rangeNoOverlapMargin[line_index] =
+        EstimateIndexRange(spectralAxis, mu, lambdaRange, winsize);
+    ;
+  }
 }
 
 void CLineModelElement::EstimateLineVisbility(
     Int32 line_index, const CSpectrumSpectralAxis &spectralAxis,
-    const TInt32Range &supportRange, Float64 line_lambda, Float64 sigma,
-    Float64 max_offset) {
+    Float64 line_lambda, Float64 sigma, Float64 max_offset) {
 
-  if (supportRange.GetBegin() > supportRange.GetEnd()) {
+  auto const &supportRange = m_range[line_index];
+  if (supportRange.GetLength() < 0) {
     // in this case the line is completely outside the
     // lambdarange
     m_OutsideLambdaRangeList[line_index] = true;
@@ -217,9 +227,11 @@ void CLineModelElement::EstimateLineVisbility(
       [line_lambda](Float64 lambda) { return std::abs(lambda - line_lambda); });
   auto const &min_distance =
       *std::min_element(distance.begin(), distance.end());
-
-  if (min_distance > (m_maxDistanceToLine * sigma + max_offset) ||
-      (nsupport < m_minSamplesNumberForLineFit)) {
+  auto const maxDistanceToLine = getElementParam()->m_maxDistanceToLine;
+  auto const minSamplesNumberForLineFit =
+      getElementParam()->m_minSamplesNumberForLineFit;
+  if (min_distance > (maxDistanceToLine * sigma + max_offset) ||
+      (nsupport < minSamplesNumberForLineFit)) {
     m_OutsideLambdaRangeList[line_index] = true;
   } else {
     m_OutsideLambdaRangeList[line_index] = false;
@@ -286,16 +298,15 @@ void CLineModelElement::initSupport(const CSpectrumSpectralAxis &spectralAxis,
 
   Int32 nLines = GetSize();
   m_OutsideLambdaRange = true;
-  m_StartNoOverlap.assign(nLines, undefIdx);
-  m_EndNoOverlap.assign(nLines, undefIdx);
-  m_StartTheoretical.assign(nLines, undefIdx);
-  m_EndTheoretical.assign(nLines, undefIdx);
+  m_rangeNoOverlap.assign(nLines, TInt32Range{undefIdx, undefIdx});
+  m_rangeNoOverlapMargin.assign(nLines, TInt32Range{undefIdx, undefIdx});
+  m_range.assign(nLines, TInt32Range{undefIdx, undefIdx});
   m_OutsideLambdaRangeList.assign(nLines, true);
   m_LineIsActiveOnSupport.assign(nLines, TBoolList(nLines, false));
+  m_sortedLineIndices.clear();
 
   for (Int32 index = 0; index != nLines; ++index) {
-    EstimateTheoreticalSupport(index, spectralAxis, redshift, lambdaRange,
-                               max_offset);
+    EstimateSupport(index, spectralAxis, redshift, lambdaRange, max_offset);
     // set the lines active on their own support
     m_LineIsActiveOnSupport[index][index] = true;
   }
@@ -303,44 +314,54 @@ void CLineModelElement::initSupport(const CSpectrumSpectralAxis &spectralAxis,
 }
 
 bool CLineModelElement::mergeIfOverlapping(Int32 i, Int32 j) {
-  Int32 x1 = m_StartNoOverlap[i];
-  Int32 x2 = m_EndNoOverlap[i];
-  Int32 y1 = m_StartNoOverlap[j];
-  Int32 y2 = m_EndNoOverlap[j];
 
-  Int32 max = std::max(x1, y1);
-  Int32 min = std::min(x2, y2);
-
-  if (max - min >= 0)
+  if (!m_rangeNoOverlap[i].unionWith(m_rangeNoOverlap[j]))
     return false;
 
-  m_StartNoOverlap[i] = std::min(x1, y1);
-  m_EndNoOverlap[i] = std::max(x2, y2);
-
   // deactivate j
-  m_StartNoOverlap[j] = m_EndNoOverlap[i];
-  m_EndNoOverlap[j] = m_EndNoOverlap[i] - 1;
+  m_rangeNoOverlap[j] = {m_rangeNoOverlap[i].GetEnd(),
+                         m_rangeNoOverlap[i].GetEnd() - 1};
+  if (getElementParam()->m_useAmpOffsetsCoeffs)
+    m_rangeNoOverlapMargin[j] = {m_rangeNoOverlapMargin[i].GetEnd(),
+                                 m_rangeNoOverlapMargin[i].GetEnd() - 1};
+
   return true;
 }
 
-void CLineModelElement::resolveOverlaps() {
+void CLineModelElement::sortLinesByLeftIndex() {
   Int32 nLines = GetSize();
+  m_sortedLineIndices.reserve(nLines);
 
-  auto shouldContinue = [this](Int32 index) {
-    return m_OutsideLambdaRangeList[index] ||
-           m_StartNoOverlap[index] > m_EndNoOverlap[index];
+  // list visible lines
+  for (Int32 index = 0; index != nLines; ++index)
+    if (!m_OutsideLambdaRangeList[index])
+      m_sortedLineIndices.push_back(index);
+
+  // then sort by left most index
+  std::sort(m_sortedLineIndices.begin(), m_sortedLineIndices.end(),
+            [this](Int32 l, Int32 r) {
+              return m_rangeNoOverlap[l] < m_rangeNoOverlap[r];
+            });
+}
+
+void CLineModelElement::mergeOverlapingLines() {
+  auto alreadyMerged = [this](Int32 index) {
+    return m_rangeNoOverlap[index].GetLength() < 0;
   };
 
-  for (Int32 i = 0; i != nLines; ++i) {
-    if (shouldContinue(i))
+  // assumes lines are sorted by left boundary
+  for (auto iter_left = m_sortedLineIndices.begin();
+       iter_left != m_sortedLineIndices.end(); ++iter_left) {
+    if (alreadyMerged(*iter_left))
       continue;
 
-    for (Int32 j = 0; j != nLines; ++j) {
-      if (shouldContinue(j) || i == j)
+    for (auto iter_right = iter_left + 1;
+         iter_right != m_sortedLineIndices.end(); ++iter_right) {
+      if (alreadyMerged(*iter_right))
         continue;
 
-      if (mergeIfOverlapping(i, j)) {
-        propagateOverlap(i, j);
+      if (mergeIfOverlapping(*iter_left, *iter_right)) {
+        propagateOverlap(*iter_left, *iter_right);
       }
     }
   }
@@ -351,39 +372,28 @@ void CLineModelElement::prepareSupport(
     const TFloat64Range &lambdaRange, Float64 max_offset) {
 
   initSupport(spectralAxis, redshift, lambdaRange, max_offset);
-  bool hasDuplicates = true;
-  Int32 icmpt = 0;
-  const Int32 ncmpt = 20;
 
-  while (hasDuplicates && icmpt < ncmpt) {
-    icmpt++;
-    resolveOverlaps();
-    hasDuplicates = detectDuplicateOverlaps();
+  if (!IsOutsideLambdaRange()) {
+    sortLinesByLeftIndex();
+    mergeOverlapingLines();
+    if (detectRemainingOverlaps())
+      THROWG(ErrorCode::INTERNAL_ERROR, "Remaining overlaps");
   }
 }
 
-bool CLineModelElement::detectDuplicateOverlaps() {
+bool CLineModelElement::detectRemainingOverlaps() {
   Int32 nLines = GetSize();
 
   for (Int32 i = 0; i != nLines; ++i) {
     if (m_OutsideLambdaRangeList[i])
       continue;
 
-    for (Int32 j = 0; j != nLines; ++j) {
-      if (m_OutsideLambdaRangeList[j] || i == j)
+    for (Int32 j = i + 1; j != nLines; ++j) {
+      if (m_OutsideLambdaRangeList[j])
         continue;
 
-      Int32 x1 = m_StartNoOverlap[i];
-      Int32 x2 = m_EndNoOverlap[i];
-      Int32 y1 = m_StartNoOverlap[j];
-      Int32 y2 = m_EndNoOverlap[j];
-
-      Int32 max = std::max(x1, y1);
-      Int32 min = std::min(x2, y2);
-
-      if (max - min < 0) {
+      if (m_rangeNoOverlap[i].HasIntersectionWith(m_rangeNoOverlap[j]))
         return true;
-      }
     }
   }
   return false;
@@ -421,23 +431,39 @@ void CLineModelElement::propagateOverlap(Int32 i, Int32 j) {
  *m_OutsideLambdaRange, for each m_Lines element which is also not outside
  *lambda range, add its support to the return value.
  **/
+TInt32RangeList
+CLineModelElement::getSupportNoOverlap(bool polynomialMargin) const {
+  TInt32RangeList support;
+  if (m_OutsideLambdaRange)
+    return support;
+
+  auto const &rangeNoOverlap =
+      polynomialMargin ? m_rangeNoOverlapMargin : m_rangeNoOverlap;
+  for (Int32 index = 0; index != GetSize(); ++index) {
+    if (m_OutsideLambdaRangeList[index])
+      continue;
+
+    support.push_back(rangeNoOverlap[index]);
+  }
+  return support;
+}
+
+TInt32RangeList CLineModelElement::getSortedSupportNoOverlap() const {
+  TInt32RangeList support;
+  if (m_OutsideLambdaRange)
+    return support;
+
+  ASSERT(!m_sortedLineIndices.empty(), ErrorCode::INTERNAL_ERROR,
+         "m_sortedLineIndices was not computed");
+
+  for (Int32 index : m_sortedLineIndices)
+    support.push_back(m_rangeNoOverlap[index]);
+
+  return support;
+}
+
 TInt32RangeList CLineModelElement::getSupport() const {
   TInt32RangeList support;
-  if (m_OutsideLambdaRange)
-    return support;
-
-  for (Int32 index = 0; index != GetSize(); ++index) {
-    if (m_OutsideLambdaRangeList[index])
-      continue;
-
-    support.push_back(
-        TInt32Range(m_StartNoOverlap[index], m_EndNoOverlap[index]));
-  }
-  return support;
-}
-
-TInt32RangeList CLineModelElement::getTheoreticalSupport() const {
-  TInt32RangeList support;
 
   if (m_OutsideLambdaRange)
     return support;
@@ -446,31 +472,9 @@ TInt32RangeList CLineModelElement::getTheoreticalSupport() const {
     if (m_OutsideLambdaRangeList[index])
       continue;
 
-    support.push_back(
-        TInt32Range(m_StartTheoretical[index], m_EndTheoretical[index]));
+    support.push_back(m_range[index]);
   }
 
-  return support;
-}
-
-/**
- * \brief Creates an empty list of ranges as the return value. If not
- *m_OutsideLambdaRange, for each m_Lines element belonging to the argument
- *subeIdx which is also not outside lambda range, add its support to the return
- *value.
- **/
-TInt32Range CLineModelElement::getSupportSubElt(Int32 index) const {
-  TInt32Range support =
-      TInt32Range(m_StartNoOverlap[index], m_EndNoOverlap[index]);
-  return support;
-}
-
-/**
- * \brief Returns the theoretical support of the line (sub-element).
- **/
-TInt32Range CLineModelElement::getTheoreticalSupportSubElt(Int32 index) const {
-  TInt32Range support =
-      TInt32Range(m_StartTheoretical[index], m_EndTheoretical[index]);
   return support;
 }
 
@@ -568,7 +572,7 @@ void CLineModelElement::addToSpectrumModel(
     if (line_index != undefIdx && !(m_LineIsActiveOnSupport[line_index][index]))
       continue;
 
-    for (Int32 i = m_StartNoOverlap[index]; i <= m_EndNoOverlap[index]; i++) {
+    for (Int32 i : m_rangeNoOverlap[index]) {
       Float64 lambda = modelspectralAxis[i];
       Float64 Yi =
           getModelAtLambda(lambda, redshift, continuumfluxAxis[i], index);
@@ -578,9 +582,8 @@ void CLineModelElement::addToSpectrumModel(
                Formatter() << "NaN flux Line: "
                            << getElementParam()->GetLineName(index)
                            << ", ContinuumFlux " << continuumfluxAxis[i]
-                           << ", ModelAtLambda Yi = " << Yi << " for range ["
-                           << m_StartNoOverlap[index] << ", "
-                           << m_EndNoOverlap[index] << "]");
+                           << ", ModelAtLambda Yi = " << Yi << "for range "
+                           << m_rangeNoOverlap[index]);
     }
   }
   return;
@@ -605,7 +608,7 @@ void CLineModelElement::addToSpectrumModelDerivVel(
     if (std::isnan(A))
       THROWG(ErrorCode::INTERNAL_ERROR, "FittedAmplitude cannot be NAN");
 
-    for (Int32 i = m_StartNoOverlap[index]; i <= m_EndNoOverlap[index]; i++) {
+    for (Int32 i : m_rangeNoOverlap[index]) {
 
       Float64 const x = modelspectralAxis[i];
       auto const &[mu, sigma] =
@@ -809,7 +812,7 @@ void CLineModelElement::initSpectrumModel(
     if (line_index != undefIdx && !(m_LineIsActiveOnSupport[line_index][index]))
       continue;
 
-    for (Int32 i = m_StartNoOverlap[index]; i <= m_EndNoOverlap[index]; i++)
+    for (Int32 i : m_rangeNoOverlap[index])
       modelfluxAxis[i] = continuumfluxAxis[i];
   }
   return;
@@ -832,7 +835,7 @@ void CLineModelElement::initSpectrumModelPolynomial(
     if (line_index != undefIdx && !(m_LineIsActiveOnSupport[line_index][index]))
       continue;
 
-    for (Int32 i = m_StartNoOverlap[index]; i <= m_EndNoOverlap[index]; i++)
+    for (Int32 i : m_rangeNoOverlap[index])
       modelfluxAxis[i] =
           m_ElementParam->m_ampOffsetsCoeffs.getValue(spcAxis[i]);
   }
@@ -870,9 +873,9 @@ void CLineModelElement::dumpElement(std::ostream &os) const {
        << m_ElementParam->m_SignFactors[i] << "\t "
        << m_ElementParam->m_FittedAmplitudes[i] << "\t"
        << m_ElementParam->m_FittedAmplitudesStd[i] << "\t"
-       << m_ElementParam->m_NominalAmplitudes[i] << "\t" << m_StartNoOverlap[i]
-       << "\t" << m_EndNoOverlap[i] << "\t" << m_StartTheoretical[i] << "\t"
-       << m_EndTheoretical[i] << "\n";
+       << m_ElementParam->m_NominalAmplitudes[i] << "\t"
+       << m_rangeNoOverlap[i].GetBegin() << "\t" << m_rangeNoOverlap[i].GetEnd()
+       << "\t" << m_range[i].GetBegin() << "\t" << m_range[i].GetEnd() << "\n";
   }
 
   os << "\n";
@@ -903,10 +906,7 @@ Int32 CLineModelElement::computeCrossProducts(
     if (line_index != undefIdx && !isLineActiveOnSupport(line_index, index))
       continue;
 
-    //    Log.LogDebug(Formatter()<<"redshift="<<redshift<<"
-    //    line_index="<<line_index<<
-    //    "start="<<getStartNoOverlap(index)<<"end="<<getEndNoOverlap(index));
-    for (Int32 i = getStartNoOverlap(index); i <= getEndNoOverlap(index); i++) {
+    for (Int32 i : m_rangeNoOverlap[index]) {
       c = continuumfluxAxis[i];
       y = noContinuumfluxAxis[i];
       x = spectralAxis[i];